+++ packages/shipit/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @ship-it/shipit
+# @ship-it-ui/shipit
 
 ## 0.0.1
 
@@ -6,5 +6,5 @@
 
 - 1035968: v0 launch
 - Updated dependencies [1035968]
-  - @ship-it/icons@0.0.1
-  - @ship-it/ui@0.0.1
+  - @ship-it-ui/icons@0.0.1
+  - @ship-it-ui/ui@0.0.1